+++ app/helpers/application_helper.rb
@@ -52,6 +52,10 @@ module ApplicationHelper
 		if (bool) then return :true else return :false end
 	end
 
+	def amount_to_f(amount)
+	  amount / 100.00
+	end
+
 
 	private
 	## inspired by gems/actionview-4.2.5/lib/action_view/helpers/translation_helper.rb