+++ config/routes.rb
@@ -48,6 +48,11 @@ Rails.application.routes.draw do
     resources :notifications, only: [:index, :show, :update] do
       match :update_all, path: '/', via: [:put, :patch], on: :collection
     end
+    resources :wallet, only: [] do
+      get '/by_user/:user_id', action: 'by_user', on: :collection
+      get :transactions, on: :member
+      put :credit, on: :member
+    end
 
     # for homepage
     get '/last_subscribed/:last' => "members#last_subscribed"